키노출제거

+++ upload_aws.py
@@ -1,14 +1,14 @@
 import boto3
 from botocore.exceptions import NoCredentialsError
 from botocore.exceptions import NoCredentialsError, PartialCredentialsError
+from enc_dec import *
+from db_insert import get_aws
 
-aws_access_key_id = ''
-aws_secret_access_key = ''
 def upload_s3_object(bucket_name, object_key, local_file_path, flag):
     # AWS S3 자격 증명 설정
-    
+    a = get_aws()
     # S3 클라이언트 생성
-    s3 = boto3.client('s3', aws_access_key_id=aws_access_key_id, aws_secret_access_key=aws_secret_access_key)
+    s3 = boto3.client('s3', aws_access_key_id=a[0], aws_secret_access_key=a[1])
     
     try:
         if flag =='D':
@@ -27,16 +27,17 @@ def upload_s3_object(bucket_name, object_key, local_file_path, flag):
         print(f'파일 업로드 중 오류 발생: {str(e)}')
     
 def check_s3_object_exists(bucket_name, object_key):
+    a = get_aws()
+    
     # AWS S3 자격 증명 설정
     #bucket_name = 'hjw7603'
     
     # S3 클라이언트 생성
-    s3 = boto3.client('s3', aws_access_key_id=aws_access_key_id, aws_secret_access_key=aws_secret_access_key)
+    s3 = boto3.client('s3', aws_access_key_id=a[0], aws_secret_access_key=a[1])
     try:
         # AWS 리전 및 인증 설정
         #s3 = boto3.client('s3')
-        
-        #bucket_name = 'hjw7603'
+
         # S3 버킷에서 객체 가져오기 시도
         s3.head_object(Bucket=bucket_name, Key=object_key)
 
@@ -55,6 +56,7 @@ def check_s3_object_exists(bucket_name, object_key):
         else:
             print(f"오류 발생: {e}")
             return False
+
 def upload_file_aws(bucket_name, object_key, local_file_path):
     # 체크할 S3 버킷 및 객체 키 설정
     if check_s3_object_exists(bucket_name, object_key):
@@ -64,29 +66,31 @@ def upload_file_aws(bucket_name, object_key, local_file_path):
     else:
         print(f"{object_key} 파일이 {bucket_name} 버킷에 존재하지 않습니다.")
         upload_s3_object(bucket_name, object_key, local_file_path,'')
+
 def delete_s3_object(bucket_name, object_key):
     # AWS S3 자격 증명 설정
-   # bucket_name = 'hjw7603'
-    
+    a = get_aws()
     # S3 클라이언트 생성
-    s3 = boto3.client('s3', aws_access_key_id=aws_access_key_id, aws_secret_access_key=aws_secret_access_key)
-        
     try:
-        s3.delete_object(Bucket=bucket_name, Key=object_key)
-        print(f"{object_key} 파일이 {bucket_name} 버킷에서 삭제되었습니다.")
+        s3 = boto3.client('s3', aws_access_key_id=a[0], aws_secret_access_key=a[1])
+        if check_s3_object_exists(bucket_name, object_key):
+            s3.delete_object(Bucket=bucket_name, Key=object_key)
+            print(f"{object_key} 파일이 {bucket_name} 버킷에서 삭제되었습니다.")
+        else:
+            print(f"{object_key} 파일이 {bucket_name} 버킷에 존재하지 않습니다.")
             
-
     #except FileNotFoundError:
-        #ㄴprint(f'{object_key} 파일을 찾을 수 없습니다.')
+        #print(f'{object_key} 파일을 찾을 수 없습니다.')
     except NoCredentialsError:
         print('AWS 자격 증명이 설정되지 않았습니다.')
     except Exception as e:
-        print(f'파일 삭제중 오류 발생: {str(e)}')        
+        print(f'파일 삭제중 오류 발생: {str(e)}')  
+
 def delete_file_aws(bucket_name, object_key):
     # S3 객체의 존재 여부 확인
-    #if check_s3_object_exists(bucket_name, object_key):
-    delete_s3_object(bucket_name, object_key)
+    if check_s3_object_exists(bucket_name, object_key):
+        delete_s3_object(bucket_name, object_key)
     
 if __name__ == "__main__":
     # 체크할 S3 버킷 및 객체 키 설정
-    print("main")
+    print(delete_s3_object("hjw7603", "share/hjw603.txt"))
